--- a/serverless/apis/recommend/requirements.txt
+++ b/serverless/apis/recommend/requirements.txt
@@ -1,6 +1,4 @@
-opensearch-py==2.7.1
-requests-aws4auth==1.3.1
-boto3==1.35.57
+elasticsearch==8.7.0
 google-cloud-bigquery==3.11.4
 aws-xray-sdk==2.14.0
 pandas==2.0.3