--- a/serverless/crawler/config/config.go
+++ b/serverless/crawler/config/config.go
@@ -1,20 +1,21 @@
 package config
 
 import (
-	"context"
+	"crypto/tls"
+	"fmt"
 	"log"
+	"net/http"
 	"os"
+	"strconv"
 	"sync"
 
-	awsConfig "github.com/aws/aws-sdk-go-v2/config"
-	"github.com/opensearch-project/opensearch-go/v2"
-	"github.com/opensearch-project/opensearch-go/v2/signer/awsv2"
+	"github.com/elastic/go-elasticsearch/v8"
 )
 
 type Config struct {
 	CrawlerConfigFilePath string
 	IndexName             string
-	OpenSearchConfig      opensearch.Config
+	ElasticSearchConfig   elasticsearch.Config
 	S3BucketName          string
 	KMSKeyARN             string
 	TextHandlerLambdaName string
@@ -30,7 +31,7 @@ func GetConfigSingletonInstance() *Config {
 		instance = &Config{
 			CrawlerConfigFilePath: os.Getenv("CRAWLER_CONFIG_FILE_PATH"),
 			IndexName:             os.Getenv("INDEX_NAME"),
-			OpenSearchConfig:      getOpenSearchConfig(os.Getenv("OPENSEARCH_ENDPOINT")),
+			ElasticSearchConfig:   getElasticsearchConfig(),
 			S3BucketName:          os.Getenv("S3_BUCKET_NAME"),
 			KMSKeyARN:             os.Getenv("KMS_KEY_ARN"),
 			TextHandlerLambdaName: os.Getenv("TEXT_HANDLER_LAMBDA_NAME"),
@@ -39,21 +40,25 @@ func GetConfigSingletonInstance() *Config {
 	return instance
 }
 
-func getOpenSearchConfig(endpointUrl string) opensearch.Config {
-	// AWS 설정 로드
-	cfg, err := awsConfig.LoadDefaultConfig(context.TODO(), awsConfig.WithRegion("ap-northeast-2"))
+func getElasticsearchConfig() elasticsearch.Config {
+	host := os.Getenv("ELASTICSEARCH_HOST")
+	portStr := os.Getenv("ELASTICSEARCH_PORT")
+	username := os.Getenv("ELASTICSEARCH_USERNAME")
+	password := os.Getenv("ELASTICSEARCH_PASSWORD")
+
+	port, err := strconv.Atoi(portStr)
 	if err != nil {
-		log.Fatal("Failed to load AWS config: " + err.Error())
+		log.Fatalf("Invalid port number: %s", err)
 	}
 
-	// AWS 서명자 생성
-	signer, err := awsv2.NewSignerWithService(cfg, "es")
-	if err != nil {
-		log.Fatal("Failed to create signer: " + err.Error())
-	}
-
-	return opensearch.Config{
-		Addresses: []string{endpointUrl},
-		Signer:    signer,
+	return elasticsearch.Config{
+		Addresses: []string{fmt.Sprintf("https://%s:%d", host, port)},
+		Username:  username,
+		Password:  password,
+		Transport: &http.Transport{
+			TLSClientConfig: &tls.Config{
+				InsecureSkipVerify: true,
+			},
+		},
 	}
 }