… serverless/apis/search/go.sum
… serverless/crawler/go.sum