## #️⃣ Related Issues
- This closes #48

## 📝 Work Details
- Migrated search infrastructure from AWS OpenSearch Free Tier to self-hosted Elasticsearch on EC2 using Docker
- Updated Lambda functions (recommend API, crawler, search API) to use Elasticsearch client instead of OpenSearch
- Modified query construction and result parsing to align with Elasticsearch syntax
- Achieved significant performance improvement: reduced query time from 1.5s to 362ms (76% reduction)

### Performance Comparison
OpenSearch Free Tier Instance
![image](https://github.com/user-attachments/assets/3f19939d-f82b-4dce-a4f4-045b4a56c000)

Self-hosted Elasticsearch on EC2 Free Tier Instance(with additional 4GiB swap memory, 20GiB storage) 
![image](https://github.com/user-attachments/assets/fc8ac45c-5255-4e3e-ba05-52411498a056)

Key metrics to highlight in the APM screenshots:
- Before: Average query time of 1.5 seconds with OpenSearch
- After: Average query time of 362ms with Elasticsearch

### Search API Response Time
![image](https://github.com/user-attachments/assets/38b82959-cebc-453c-8438-b126d4f0be0d)
Total Search API response time: 1175ms
- Lambda cold start time: 780ms
- Actual Lambda execution time: 395ms